Now creates missing directories for json and images

+++ fetch_imgs.py
@@ -7,6 +7,8 @@ import common
 IMG_DIR = "images/" # must end in "/"
 
 def update_json(start_sol, end_sol):
+  if not os.path.isdir(common.JSON_DIR):
+    os.makedirs(common.JSON_DIR)
   common.download_file(common.MANIFEST_PATH, "http://mars.jpl.nasa.gov/msl-raw-images/image/image_manifest.json")
   
   manifest = common.load_manifest()
@@ -17,12 +19,15 @@ def update_json(start_sol, end_sol):
       common.download_file(outfile, url)
 
 def clear_downloaded_manifests():
-  for filename in os.listdir(common.JSON_DIR):
-    os.remove(common.JSON_DIR + filename)
+  if os.path.isdir(common.JSON_DIR):
+    for filename in os.listdir(common.JSON_DIR):
+      os.remove(common.JSON_DIR + filename)
 
 def get_full_images(start_sol, end_sol, instrument_startswith):
   clear_downloaded_manifests()
   update_json(start_sol, end_sol)
+  if not os.path.isdir(IMG_DIR):
+    os.makedirs(IMG_DIR)
   sol_dirs = os.listdir(IMG_DIR)
   for image_list in common.load_image_lists():
     sol = image_list["sol"]
@@ -42,5 +47,8 @@ def get_full_images(start_sol, end_sol, instrument_startswith):
         else:
           common.download_file(local_path, url)
 
+# Makes an index for use by the javascript
+def make_index_of_downloaded_photos():
+  pass
 
 get_full_images(0, 10, "NAV_")
